Initial import of rust version of mitsusplit
ESP32-C3 firmware implementing a web/mqtt interface to mitsubishi
mini-split air handlers.

Basic outline of a runnable app that, if unprovisioned, boots up in
provisioning mode, appearing as a wifi access point to configure the
local wifi network to connect to (although the configuration interface
is not yet in place). If provisioned, it boots up in client-only mode to
host a webapp and mqtt.

Remaining to do are:

1. Complete the provisioning web UI.

2. Complete the full configuration web UI.

3. Implement serial interface to AC handler for comfort control

4. Implement mqtt gateway to query/configure comfort settings